kernel: move float operations out of thread.c
Move float operation out and add missing vrfy hook for enabling float.
